Provide type for external events (i/o) SpeechState
