+++ src/speechstate.ts
@@ -8,7 +8,7 @@ import {
 import { ttsMachine } from "./tts";
 import { asrMachine } from "./asr";
 
-import { Settings, Agenda, Hypothesis, RecogniseParameters } from "./types";
+import { Settings, SpeechStateEvent } from "./types";
 interface SSContext {
   settings: Settings;
   audioContext?: AudioContext;
@@ -16,36 +16,12 @@ interface SSContext {
   ttsRef?: any;
 }
 
-/** events sent to the spawned `speechstate` machine **/
-type SSEventExtIn =
-  | { type: "PREPARE" }
-  | { type: "CONTROL" }
-  | { type: "STOP" }
-  | { type: "SPEAK"; value: Agenda }
-  | { type: "LISTEN"; value: RecogniseParameters };
-
-/** for sendParent, not type-checked */
-type SSEventExtOut =
-  | { type: "ASR_NOINPUT" }
-  | { type: "ASRTTS_READY" }
-  | { type: "ASR_STARTED" }
-  | { type: "TTS_STARTED" }
-  | { type: "SPEAK_COMPLETE" }
-  | { type: "RECOGNISED"; value: Hypothesis[]; nluValue?: any };
-
-type SSEventIntIn =
-  | { type: "TTS_READY" }
-  | { type: "ASR_READY" }
-  | { type: "TTS_ERROR" };
-
-type SSEvent = SSEventIntIn | SSEventExtIn | SSEventExtOut;
-
 const speechstate = createMachine(
   {
     types: {} as {
       input: Settings;
       context: SSContext;
-      events: SSEvent;
+      events: SpeechStateEvent;
     },
     context: ({ input }) => ({
       settings: input,
